% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculateModel}
\alias{calculateModel}
\title{Build the model with the input data}
\usage{
calculateModel(root.dir = getwd(), cache = TRUE, wof)
}
\arguments{
\item{root.dir}{Workspace root directory used to load cached data}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{wof}{If it is not NULL, indicates which feature is going to calculate by current computer.}
}
\value{
the trained model
}
\description{
Build the model with the input data
}
