% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluates the model with the input data (score each packet)}
\usage{
evaluateModel(model.ds, root.dir = getwd(), cache = TRUE, wof = NULL)
}
\arguments{
\item{root.dir}{Workspace root directory used to load cached data}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{wof}{If it is not NULL, indicates which feature is going to calculate by current computer.}
}
\value{
the trained model
}
\description{
Evaluates the model with the input data (score each packet)
}
