% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{getLabeledTestingData}
\alias{getLabeledTestingData}
\title{Just adds precalculated labels columns to testing data}
\usage{
getLabeledTestingData(testing.raw.ds, label.testing.raw.ds)
}
\arguments{
\item{testing.raw.ds}{Data set for evaluation.}

\item{label.testing.raw.ds}{Precalculated label data}
}
\value{
testing data with label columns
}
\description{
Just adds precalculated labels columns to testing data
}
