% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{getScoredTestingData}
\alias{getScoredTestingData}
\title{Scales scoring values between 1 and 100}
\usage{
getScoredTestingData(testing.scored.ds)
}
\arguments{
\item{testing.scored.ds}{scored dataset without scaling}
}
\value{
testing dataset with scoring values
}
\description{
Scales scoring values between 1 and 100
}
