% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{labelTestingDS}
\alias{labelTestingDS}
\title{Creates a label index, it says for each testing entry position if it is really or not an attack
and related attack data (name, attackid and type)}
\usage{
labelTestingDS(testing.raw.ds, label.attacklist.raw.ds,
  label.attacktypes.raw.ds, cache = FALSE, root.dir = getwd())
}
\arguments{
\item{testing.raw.ds}{testing data}

\item{label.attacklist.raw.ds}{attack label data to figure out if from entry in the testing data is or not an attack}

\item{label.attacktypes.raw.ds}{attack type information. Contains a pair (attack nama, attack type)}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{root.dir}{Workspace root directory used to load cached data}
}
\value{
label dataset
}
\description{
Creates a label index, it says for each testing entry position if it is really or not an attack
and related attack data (name, attackid and type)
}
