% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{measureResults}
\alias{measureResults}
\title{Extract measures from the evaluation results}
\usage{
measureResults(testing.raw.scored.ds, root.dir = getwd(), cache = TRUE)
}
\arguments{
\item{root.dir}{Workspace root directory used to load cached data}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{wof}{If it is not NULL, indicates which feature is going to calculate by current computer.}
}
\value{
the trained model
}
\description{
Extract measures from the evaluation results
}
