% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{scoreFeature}
\alias{scoreFeature}
\title{Calculates score for a specific feature}
\usage{
scoreFeature(model.ds, testing.features.ds, testing.timestamps.ds, nr,
  indexFeature, initialTs)
}
\arguments{
\item{model.ds}{Model used to decide whether a register is an attack or not}

\item{testing.features.ds}{Data set for evaluation. Decide if some packet is an attack or not}

\item{testing.timestamps.ds}{Data set with timestamps.}

\item{nr}{Precalculated n/r value}

\item{indexFeature}{Position (column) of the current feature}

\item{initialTs}{Initial timestamp in order to calculate the time since last anomaly}
}
\value{
the scoring for a specific feature
}
\description{
Calculates score for a specific feature
}
