% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{scoring}
\alias{scoring}
\title{For each testing entry calculates a scoring. It says if current entry is or not an attack.}
\usage{
scoring(model.ds, testing.raw.ds, root.dir = getwd(), cache = TRUE,
  wof = NULL)
}
\arguments{
\item{model.ds}{Model used to decide whether a register is an attack or not}

\item{testing.raw.ds}{Data set for evaluation. Decide if some packet is an attack or not}

\item{root.dir}{Workspace root directory used to load cached data}

\item{cache}{Do not do any calculation, use precalculated data}

\item{wof}{Indicates which feature is going to calculate}
}
\description{
For each testing entry calculates a scoring. It says if current entry is or not an attack.
}
