% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{train}
\alias{train}
\title{Train the model with input data}
\usage{
train(training.raw.ds, cache = FALSE, root.dir = getwd(), wof = NULL)
}
\arguments{
\item{training.raw.ds}{Input data}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{root.dir}{Workspace root directory used to load cached data}

\item{wof}{If it is not NULL, indicates which feature is going to calculate by current computer.}
}
\value{
the trained model
}
\description{
Train the model with input data
}
