% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phad-c32.R
\name{trainFeature}
\alias{trainFeature}
\title{Train specific feature of the model with input feature data}
\usage{
trainFeature(featureObs, nFeature, cache = FALSE, root.dir = getwd())
}
\arguments{
\item{featureObs}{Observations for current feature (vector)}

\item{nFeature}{Feature position in order to save or load from cache}

\item{cache}{Do not do any calculation, use precalculated data if it is available}

\item{root.dir}{Workspace root directory used to load cached data}
}
\value{
trained model for the specific feature
}
\description{
Train specific feature of the model with input feature data
}
